+++ packages/ragbits-core/src/ragbits/core/embeddings/noop.py
@@ -1,6 +1,6 @@
 from ragbits.core.audit import traceable
 from ragbits.core.embeddings.base import Embedder
-from ragbits.core.options import Options
+from ragbits.core.options import Options, OptionsT
 
 
 class NoopEmbedder(Embedder[Options]):
@@ -14,6 +14,29 @@ class NoopEmbedder(Embedder[Options]):
 
     options_cls = Options
 
+    def __init__(
+        self,
+        default_options: OptionsT | None = None,
+        return_values: list[list[list[float]]] | None = None,
+        image_return_values: list[list[list[float]]] | None = None,
+    ) -> None:
+        """
+        Constructs a new NoopEmbedder instance.
+
+        Args:
+            default_options: The default options for the component.
+            return_values: The embeddings to return for text input. Each time the embed_text method is called,
+                the next list of embeddings is returned, after being trimmed / repeated to match the number of inputs.
+                After all return_values have been used, the cycle starts again. Default is a single vector of [0.1, 0.1]
+            image_return_values: The embeddings to return for image input. Similar to return_values, but for images.
+                If not provided, image embeddings are not supported.
+        """
+        super().__init__(default_options=default_options)
+        self.return_values = return_values or [[[0.1, 0.1]]]
+        self.image_return_values = image_return_values
+        self.return_cycle = 0
+        self.image_return_cycle = 0
+
     @traceable
     async def embed_text(self, data: list[str], options: Options | None = None) -> list[list[float]]:  # noqa: PLR6301
         """
@@ -24,7 +47,43 @@ class NoopEmbedder(Embedder[Options]):
             options: Additional settings used by the Embedder model.
 
         Returns:
-            A list of embedding vectors, where each vector
-            is a fixed value of [0.1, 0.1] for each input string.
+            A list of embedding vectors, one for each input text.
         """
-        return [[0.1, 0.1]] * len(data)
+        # Get the right values for the current cycle
+        values = self.return_values[self.return_cycle]
+
+        # Expand the values to at least match the number of inputs
+        values = values * (len(data) // len(values) + 1)
+
+        # Update the cycle counter
+        self.return_cycle = (self.return_cycle + 1) % len(self.return_values)
+
+        return values[: len(data)]
+
+    def image_support(self) -> bool:
+        """
+        Check if the model supports image embeddings, which is the case if image_return_values is provided.
+
+        Returns:
+            True if the model supports image embeddings, False otherwise.
+        """
+        return self.image_return_values is not None
+
+    @traceable
+    async def embed_image(self, images: list[bytes], options: Options | None = None) -> list[list[float]]:
+        """
+        Embeds a list of images into a list of vectors.
+
+        Args:
+            images: A list of input image bytes to embed.
+            options: Additional settings used by the Embedder model.
+
+        Returns:
+            A list of embedding vectors, one for each input image.
+        """
+        if self.image_return_values is None:
+            raise NotImplementedError("Image embeddings are not supported by this model.")
+        values = self.image_return_values[self.image_return_cycle]
+        values = values * (len(images) // len(values) + 1)
+        self.image_return_cycle = (self.image_return_cycle + 1) % len(self.image_return_values)
+        return values[: len(images)]